// src/types/ai-paths.ts
export type JsonPathEntry = {
  path: string;
  type: "object" | "array" | "value";
};